% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_share_row_wise.R
\name{calculate_share_row_wise}
\alias{calculate_share_row_wise}
\title{calculate_share_row_wise}
\usage{
calculate_share_row_wise(dat = NULL)
}
\arguments{
\item{dat}{Data frame}
}
\value{
The function returns a data set that has the information
    on the row-wise NA share
}
\description{
Function used in `overview_na` to calculate the share of NA row-wise
}
