% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Cross-sectional data for countries}
\format{
An object of class \code{"data.frame"}.
}
\usage{
data(data)
}
\description{
This is a small toy data set that comes in a cross-sectional format where the
unit of analysis is either country-year or country-year-month.
It provides artificial information for five countries (Angola, Benin, France, Rwanda, and UK)
for a time span from 1990 to 1999.
}
\examples{
data(data)
head(data)
}
\references{
This data set was artificially created for the overviewR package.
}
\keyword{datasets}
