% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_day.R
\docType{data}
\name{data_day}
\alias{data_day}
\title{Longitudinal data for test persons}
\format{
An object of class \code{"data.frame"}.
}
\usage{
data(data_day)
}
\description{
This is a small toy data set that contains information on 16 test persons
on a daily basis.
}
\examples{
data(data_day)
head(data_day)
}
\references{
This data set was artificially created for the overviewR package.
}
\keyword{datasets}
