% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_overview_heat.R
\name{.overview_heat}
\alias{.overview_heat}
\title{.overview_tab}
\usage{
.overview_heat(
  dat = NULL,
  id = NULL,
  time = NULL,
  label = FALSE,
  perc = FALSE,
  col_low = NULL,
  col_high = NULL,
  xaxis = NULL,
  yaxis = NULL,
  theme_plot = NULL,
  exp_total = NULL,
  col_names = NULL
)
}
\arguments{
\item{dat}{The data set}

\item{id}{The scope (e.g., country codes or individual IDs). The axis is
ordered in ascending order by default.}

\item{time}{The time (e.g., time periods given by years, months, ...)}

\item{label}{If TRUE (default), the total number of observations/percentages
of observations are displayed. If FALSE, it returns no labels.}

\item{perc}{If FALSE (default) plot returns the total number of observations
per time-scope-unit. If TRUE, it returns the number of observations per
time-scope-unit in percentage}

\item{col_low}{Hex color code for the lowest value (default is "#dceaf2")}

\item{col_high}{Hex color code for the lowest value (default is "#2A5773")}

\item{xaxis}{Label of your x axis ("Time frame" is default)}

\item{yaxis}{Label of your y axis ("Sample" is default)}

\item{theme_plot}{Previously generated theme}

\item{exp_total}{Expected total number of observations (i.e. maximum)
for time unit.}

\item{col_names}{The column names (containing id and time)}
}
\value{
A ggplot
}
\description{
Internal function that calculates the `overview_tab` for data.table objects
}
