% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{find_int_runs}
\alias{find_int_runs}
\title{find_int_runs}
\usage{
find_int_runs(run = NULL)
}
\arguments{
\item{run}{Variable (integer) that should be checked for consecutive values}
}
\value{
The function returns a data set
}
\description{
Function used in `ovoerview_tab` to find running integers
}
