% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_add_na_output.R
\name{overview_add_na_output}
\alias{overview_add_na_output}
\title{overview_add_na_output}
\usage{
overview_add_na_output(dat_result = NULL, dat = NULL)
}
\arguments{
\item{dat_result}{Data.frame from `overview_na`}

\item{dat}{Data frame}
}
\value{
The function returns a data set that has the information
    on the row-wise NA share
}
\description{
Function used in `overview_na` to generate a new data frame with
    na_count and percentage share of NAs for each row
}
