% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_crosstab.R
\name{overview_crosstab}
\alias{overview_crosstab}
\title{overview_crosstab}
\usage{
overview_crosstab(dat, cond1, cond2, threshold1, threshold2, id, time)
}
\arguments{
\item{dat}{Your dataset}

\item{cond1}{The variable that describes your first condition}

\item{cond2}{The variable that describes your second condition}

\item{threshold1}{A threshold above which your first condition is fulfilled}

\item{threshold2}{A threshold above which your second condition is fulfilled}

\item{id}{Your scope (e.g., country codes or individual IDs)}

\item{time}{Your time (e.g., time periods given by years, months, ...)}
}
\value{
A data frame that contains a summary of your sample and that can be easily converted to a LaTeX output
}
\description{
This function allows you sort your sample conditionally in a nice cross table. This can be helpful if you want to get a first sense of your data before digging deeper.
}
\examples{
\dontrun{
overview_crosstab(dat = dataset, cond1 = var1, cond2 = var2, threshold1 = 50, threshold2 = 70, id = country_code, time = year)
}

library(tidyverse)

#' set.seed(68163)
df_combined <- data.frame(
  # Countries
  countries  = c(
    rep("RWA", 4),
    rep("AGO", 8),
    rep("BEN", 2),
    rep("GBR", 5),
    rep("FRA", 3)
  ),
  # Time frame
  years =
    c(
      seq(1990, 1995),
      seq(1990, 1992),
      seq(1995, 1999),
      seq(1991, 1999, by = 2),
      seq(1993, 1999, by = 3)
    ),
  # GDP
  gdp =
    runif(22, 10000, 40000),
  # Population
  population =
    runif(22, 100, 50000),
  stringsAsFactors = FALSE
)

output_crosstab <- overview_crosstab(
dat = df_combined,
cond1 = gdp,
cond2 = population,
threshold1 = 30000,
threshold2 = 15000,
id = countries,
time = years
)
}
