% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_plot_absolute.R
\name{overview_plot_absolute}
\alias{overview_plot_absolute}
\title{overview_plot_absolute}
\usage{
overview_plot_absolute(
  dat_result = NULL,
  theme_plot = NULL,
  yaxis = NULL,
  xaxis = NULL
)
}
\arguments{
\item{dat_result}{Data frame}

\item{theme_plot}{Theme for the plot (pre-defined)}

\item{yaxis}{Name for yaxis}

\item{xaxis}{Name for xaxix}
}
\value{
The function returns a ggplot
}
\description{
Function used in `overview_na` to plot the absolute share of NA values
}
