% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_print.R
\name{overview_print}
\alias{overview_print}
\title{overview_print}
\usage{
overview_print(
  obj,
  title = "Time and scope of the sample",
  crosstab = FALSE,
  cond1 = "Condition 1",
  cond2 = "Condition 2",
  save_out = FALSE
)
}
\arguments{
\item{obj}{Your overview object produced by overview_tab or overview_crosstab}

\item{title}{Title of your table (default is "Time and scope of the sample")}

\item{crosstab}{Logical argument, if TRUE produces a crosstab output, default is FALSE}

\item{cond1}{Description for your first condition (character), will be ignored if crosstab is FALSE}

\item{cond2}{Description for your second condition (character), will be ignored if crosstab is FALSE}

\item{save_out}{This is an optional argument that exports your table as .tex file, default is FALSE}
}
\value{
A LaTeX output that you can copy and paste in your text document
}
\description{
This function produces a LaTeX for both overview_tab and overview_crosstab
}
\examples{
\dontrun{
overview_print(obj = overview_object, title = "Some nice title", crosstab = FALSE)
overview_print(obj = overview_ct_object, title = "Some nice title for a cross tab", crosstab = TRUE, cond1 = "Name of first condition", cond2 = "Name of second condition")
}
}
