% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_tab.R
\name{overview_tab}
\alias{overview_tab}
\title{overview_tab}
\usage{
overview_tab(dat, id, time)
}
\arguments{
\item{dat}{A data frame or data table object}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods given by years, months, ...). Time can be a character vector containing multiple time variables.}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a 'LaTeX' output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set. If a data.table object is provided, the function uses
    data.table's syntax to perform the evaluation
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)
}
