% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_tab_df.R
\name{overview_tab_df}
\alias{overview_tab_df}
\title{overview_tab_df}
\usage{
overview_tab_df(dat2 = NULL, dat = NULL, id = NULL, time = NULL)
}
\arguments{
\item{dat2}{Your data set}

\item{dat}{Your data set}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods given by years, months, ...). There are three options to add a date variable: 1) Time can be a character vector containing **one** time variable, 2) a time variable following the YYYY-MM-DD format, or 3) or a list containing multiple time variables (`time = list(year = NULL, month = NULL, day = NULL)`).}
}
\value{
A data.frame
}
\description{
Internal function that calculates the `overview_tab` for data.frame objects
}
