% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_tab_dt.R
\name{overview_tab_dt}
\alias{overview_tab_dt}
\title{overview_tab_dt}
\usage{
overview_tab_dt(dat = NULL, id = NULL, time = NULL, col_names = NULL)
}
\arguments{
\item{dat}{Your data set}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods given by years, months, ...). There are three options to add a date variable: 1) Time can be a character vector containing **one** time variable, 2) a time variable following the YYYY-MM-DD format, or 3) or a list containing multiple time variables (`time = list(year = NULL, month = NULL, day = NULL)`).}

\item{col_names}{The column names (containing id and time)}
}
\value{
A data.table
}
\description{
Internal function that calculates the `overview_tab` for data.table objects
}
