% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_heat_plot.R
\name{theme_heat_plot}
\alias{theme_heat_plot}
\title{theme_heat_plot}
\usage{
theme_heat_plot()
}
\value{
A theme for the `overview_heat` plot
}
\description{
Defines the theme for the `overview_heat` plot function
}
