% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_na_plot.R
\name{theme_na_plot}
\alias{theme_na_plot}
\title{theme_na_plot}
\usage{
theme_na_plot()
}
\value{
A theme for the `overview_na` plot
}
\description{
Defines the theme for the `overview_na` plot function
}
