% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Cross-sectional data for countries}
\format{
An object of class \code{"data.frame"}.
\describe{
 \item{ccode}{A country code (character) for the countries in the sample (Angola, Benin, France, Rwanda, and UK.)}
 \item{year}{A value between 1990 and 1999.}
 \item{month}{An abbreviation for month (character).}
 \item{gpd}{A fake value for GDP (randomly generated).}
 \item{population}{A fake value for population (randomly generated).}
}
}
\usage{
data(toydata)
}
\description{
This is a small toy data set that comes in a cross-sectional format where the
unit of analysis is either country-year or country-year-month.
It provides artificial information for five countries (Angola, Benin, France, Rwanda, and UK)
for a time span from 1990 to 1999.
}
\examples{
\dontrun{
data(toydata)
head(toydata)
}
}
\references{
This data set was artificially created for the overviewR package.
}
\keyword{datasets}
