% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_crosstab.R
\name{overview_crosstab}
\alias{overview_crosstab}
\title{overview_crosstab}
\usage{
overview_crosstab(dat, cond1, cond2, threshold1, threshold2, id, time)
}
\arguments{
\item{dat}{Your dataset}

\item{cond1}{The variable that describes your first condition}

\item{cond2}{The variable that describes your second condition}

\item{threshold1}{A threshold above which your first condition is fulfilled}

\item{threshold2}{A threshold above which your second condition is fulfilled}

\item{id}{Your scope (e.g., country codes or individual IDs)}

\item{time}{Your time (e.g., time periods given by years, months, ...)}
}
\value{
A data frame that contains a summary of your sample and that can be easily converted to a LaTeX output
}
\description{
This function allows you sort your sample conditionally in a nice cross table. This can be helpful if you want to get a first sense of your data before digging deeper. Note, if you use a data set that has multiple observations on your id-time unit, the function automatically aggregates this information using the mean.
}
\examples{
\dontrun{
overview_crosstab(dat = dataset,
                  cond1 = var1, cond2 = var2,
                  threshold1 = 50, threshold2 = 70,
                  id = country_code, time = year)

output_crosstab <- overview_crosstab(
dat = toydata,
cond1 = gdp,
cond2 = population,
threshold1 = 25000,
threshold2 = 27000,
id = ccode,
time = year)
}
}
