% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_na.R
\name{overview_na}
\alias{overview_na}
\title{overview_na}
\usage{
overview_na(dat, yaxis = "Variables", perc = TRUE)
}
\arguments{
\item{dat}{Your data set}

\item{yaxis}{Label of your y axis ("Variables" is default)}

\item{perc}{If TRUE (default) plot returns the number of NAs in percentage}
}
\value{
A ggplot figure that presents the distribution of NAs in the
    data set
}
\description{
This function plots a ggplot to visualize the distribution of
    NAs across all variables in the data set.
}
\examples{
data(toydata)
overview_na(toydata, perc = FALSE)
}
