% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_plot.R
\name{overview_plot}
\alias{overview_plot}
\title{overview_plot}
\usage{
overview_plot(dat, id, time, xaxis = "Time frame", yaxis = "Sample")
}
\arguments{
\item{dat}{Your dataset}

\item{id}{Your scope (e.g., country codes or individual IDs)}

\item{time}{Your time (e.g., time periods given by years, months, ...)}

\item{xaxis}{Label of your x axis ("Time frame" is default)}

\item{yaxis}{Label of your y axis ("Sample" is default)}
}
\value{
A ggplot figure that presents your sample information visually
}
\description{
This function plots a ggplot to visualize the distribution of
    scope objects across the time frame.
}
\examples{
data(toydata)
overview_plot(dat = toydata, id = ccode, time = year)
}
