% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_tab.R
\name{overview_tab}
\alias{overview_tab}
\title{overview_tab}
\usage{
overview_tab(dat, id, time)
}
\arguments{
\item{dat}{Your dataset}

\item{id}{Your scope (e.g., country codes or individual IDs)}

\item{time}{Your time (e.g., time periods given by years, months, ...)}
}
\value{
A data frame that contains a summary of your sample and that can be easily converted to a LaTeX output
}
\description{
This function gives you a nice overview table for time and scope conditions
}
\examples{
overview_tab(dat = dataset, id = country_code, time = year)
}
