% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_tab.R
\name{overview_tab}
\alias{overview_tab}
\title{overview_tab}
\usage{
overview_tab(dat, id, time)
}
\arguments{
\item{dat}{Your dataset}

\item{id}{Your scope (e.g., country codes or individual IDs)}

\item{time}{Your time (e.g., time periods given by years, months, ...)}
}
\value{
A data frame that contains a summary of your sample and that can be later easily converted to a LaTeX output
}
\description{
This function gives you a nice overview table for time and scope conditions
}
\examples{
\dontrun{
overview_tab(dat = dataset, id = country_code, time = year)
}

library(tidyverse)

set.seed(68163)
df_combined <- data.frame(
  # Countries
  countries  = c(
    rep("RWA", 4),
    rep("AGO", 8),
    rep("BEN", 2),
    rep("GBR", 5),
    rep("FRA", 3)
  ),
  # Time frame
  years =
    c(
      seq(1990, 1995),
      seq(1990, 1992),
      seq(1995, 1999),
      seq(1991, 1999, by = 2),
      seq(1993, 1999, by = 3)
    ),
  # GDP
  gdp =
    runif(22, 10000, 40000),
  # Population
  population =
    runif(22, 100, 50000),
  stringsAsFactors = FALSE
)

output_table <- overview_tab(dat = df_combined, id = countries, time = years)
}
